feat: add db table

+++ Sprint03/src/adapter/output/database/CustomerRepository.cs
@@ -1,37 +1,18 @@
-﻿using System;
-using System.Linq;
-using Sprint03.domain.model;
+﻿using Sprint03.domain.model;
 using Sprint03.domain.repository;
 
 namespace Sprint03.adapter.output.database
 {
     public class CustomerRepository : ICustomerRepository
     {
-        private static CustomerRepository _instance;
-        private static readonly object _lock = new object();
         private readonly ApplicationDbContext _context;
 
-        private CustomerRepository(ApplicationDbContext context)
+        // Construtor público que aceita o ApplicationDbContext
+        public CustomerRepository(ApplicationDbContext context)
         {
             _context = context;
         }
 
-        public static CustomerRepository GetInstance(ApplicationDbContext context)
-        {
-            if (_instance == null)
-            {
-                lock (_lock)
-                {
-                    if (_instance == null)
-                    {
-                        _instance = new CustomerRepository(context);
-                    }
-                }
-            }
-            return _instance;
-        }
-
-        // Implementação dos métodos do repositório
         public Customer FindById(string id)
         {
             return _context.Customers.FirstOrDefault(c => c.Id == id);
@@ -47,16 +28,19 @@ namespace Sprint03.adapter.output.database
         {
             var existingCustomer = _context.Customers.FirstOrDefault(c => c.Id == id);
 
-            existingCustomer.Name = customer.Name;
-            existingCustomer.Phone = customer.Phone;
-            existingCustomer.Email = customer.Email;
-            existingCustomer.BirthDate = customer.BirthDate;
-            existingCustomer.Document = customer.Document;
-            existingCustomer.Cep = customer.Cep;
-            existingCustomer.AgreementId = customer.AgreementId;
+            if (existingCustomer != null)
+            {
+                existingCustomer.Name = customer.Name;
+                existingCustomer.Phone = customer.Phone;
+                existingCustomer.Email = customer.Email;
+                existingCustomer.BirthDate = customer.BirthDate;
+                existingCustomer.Document = customer.Document;
+                existingCustomer.Cep = customer.Cep;
+                existingCustomer.AgreementId = customer.AgreementId;
 
-            _context.Customers.Update(existingCustomer);
-            _context.SaveChanges();
+                _context.Customers.Update(existingCustomer);
+                _context.SaveChanges();
+            }
 
             return existingCustomer;
         }
@@ -65,8 +49,11 @@ namespace Sprint03.adapter.output.database
         {
             var customer = _context.Customers.FirstOrDefault(c => c.Id == id);
 
-            _context.Customers.Remove(customer);
-            _context.SaveChanges();
+            if (customer != null)
+            {
+                _context.Customers.Remove(customer);
+                _context.SaveChanges();
+            }
         }
     }
 }
